+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # **1.7.0**
-* Connect logging to Application Insights. Hopefully now I'll be able to see when/why the bot is down for one shard but not others.
-* Add a heartbeat message from each shard. Messages me each hour to confirm the shard is up and connected.
+* Rework the structure of logging.
+* Tweak the launch method in the README. Should be easier to develop on. I suggest VSCode.
+* Connect logging to Application Insights. Improve logging (including source shard and human readable event type).
+* No messages, message IDs, user IDs, Guild IDs, etc are logged.
+* Add a heartbeat from each shard. Logs some metrics each hour to confirm the shard is up and connected.
 
 # **1.6.0**
 * Make bot deferral mechanism work on a per-channel basis, rather than per-server.